Add all CRUD methods to App CLI.
